+++ Controllers/LoanControllers.cs
@@ -1,4 +1,6 @@
-﻿using Homework2.DataAccess;
+﻿using AutoMapper;
+using Homework2.DataAccess;
+using Homework2.DTOs;
 using Homework2.Entities;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -8,35 +10,49 @@ namespace Homework2.Controllers
     [ApiController]      //Normally I use [Controller], but in this code I prefer to use [ApiController]
                          // because with it I don't need to specify where the Post came from
     [Route("Api/Loans")]
-    public class LoanControllers: ControllerBase
+    public class LoanControllers: BaseController<Loan,LoanDTO>
     {
-        public readonly DataDbContext _DbContext;
-        public LoanControllers(DataDbContext _DbContext)
+        public LoanControllers(ApplicationDbContext context, ILogger<LoanControllers> logger, IMapper mapper)
+            :base(mapper,logger,context)
+        {
+        }
+        /*
+        public readonly ApplicationDbContext _DbContext;
+        public readonly ILogger<LoanControllers> _logger;
+        public readonly IMapper _mapper;
+        public LoanControllers(ApplicationDbContext _DbContext, ILogger<LoanControllers> logger, IMapper mapper)
         {
             this._DbContext = _DbContext;
+            this._logger = logger;
+            this._mapper = mapper;
         }
 
 
 
         [HttpGet] 
-        public async Task<List<Loan>> Gets()
+        public async Task<IEnumerable<LoanDTO>> Gets()
         {
-            return await _DbContext.Loans.ToListAsync<Loan>();
+            var loans = await _DbContext.Loans.ToListAsync<Loan>();
+
+            var loanDTOs = _mapper.Map<IEnumerable<LoanDTO>>(loans);
+            return loanDTOs;
         }
 
 
 
-        [HttpGet("{Id:int}")]   //A Get for specify Loan with Api/Loans/#x
-        public async Task<ActionResult<Loan>> Get(int id)
+        [HttpGet("{Id:int}", Name = "GetLoan")]   //A Get for specify Loan with Api/Loans/#x
+        public async Task<ActionResult<LoanDTO>> Get(int id)
         {
             var loan = await _DbContext.Loans.FirstOrDefaultAsync(l => l.Id == id);
 
-            if (loan is null)
+            var loanDTO = _mapper.Map<LoanDTO>(loan); 
+
+            if (loanDTO is null)
             {
-                return BadRequest("Loan doesn't exist");
+                return BadRequest("Loan doesn't exist");//401
             }
 
-            return loan;
+            return loanDTO;
         }
 
         [HttpPost]
@@ -44,7 +60,8 @@ namespace Homework2.Controllers
         {
             _DbContext.Add(loan);
             await _DbContext.SaveChangesAsync();
-            return Ok("Loan Creado");
+            _logger.LogInformation("Loan Created");
+            return CreatedAtRoute("GetLoan", new {id = loan.Id}, loan);//201
         }
 
 
@@ -53,25 +70,29 @@ namespace Homework2.Controllers
         {
             if (id != loan.Id)
             {
-                return BadRequest("The Ids not match");
+                return BadRequest("The Ids not match");//401
             }
 
             _DbContext.Update(loan);
             await _DbContext.SaveChangesAsync();
-            return Ok("Update Done Corretly");
+            _logger.LogInformation("Update Loan");
+            return CreatedAtRoute("GetLoan", new {id = loan.Id }, loan);//201
         }
 
         [HttpDelete("{Id:int}")] //A Delete for specify Loan with Api/Loans/#x
         public async Task<ActionResult> Delete(int id)
         {
-            var result = await _DbContext.Loans.Where(a => a.Id == id).ExecuteDeleteAsync();
+            var loan = _DbContext.Loans.FirstOrDefaultAsync(x => x.Id == id);
 
-            if (result == 0)
+            if (loan is null)
             {
-                return NotFound("Not Found some Loan With Id");
+                return NotFound("Not Found some Loan With Id");//404
             }
 
-            return Ok("Loan Deleted Correctly");
+            var result = await _DbContext.Loans.Where(a => a.Id == id).ExecuteDeleteAsync();
+            _logger.LogInformation("Deleted Loan");
+            return CreatedAtRoute("GetLoan", new {id = loan.Id}, loan);//201
         }
+        */
     }
 }
